Split core audio and embeddings modules
